// src/types/globalEnums.d.ts
export enum ResponseCode {
  OK = 0,
  CREATED = 201,
  ACCEPTED = 202,
  NO_CONTENT = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  INTERNAL_SERVER_ERROR = 500,
}

export enum ResponseMessage {
  OK = 'OK',
  CANCEL = 'CANCEL',
  ERROR = 'ERROR',
}